Code clean-up compilation warnings are removed
git-svn-id: https://svn.apache.org/repos/asf/openmeetings/trunk/singlewebapp@1582407 13f79535-47bb-0310-9956-ffa450edef68

+++ src/screenshare/java/org/apache/openmeetings/screen/webstart/CaptureScreen.java
@@ -123,7 +123,7 @@ final class CaptureScreen extends Thread {
 					if (f != null) {
 						try {
 							timestamp += timeBetweenFrames;
-							pushVideo(f, (int)timestamp);
+							pushVideo(f, timestamp);
 							if (log.isTraceEnabled()) {
 								log.trace("Sending video, timestamp: " + timestamp);
 							}
